計算系を括りだし。
